// src/app/noticias/models/noticias.ts
export interface Noticias {

  id: number;
  titulo: string;
  texto: string;
  autor: string;
  imagen: string;
  fecha: string
  eliminada: boolean;
  programa: string;
  importancia: string;
  lugar: string;
  diurna: boolean;
  nocturna: boolean;
  evento: boolean;
  noticia: boolean;
  favorita: boolean;



}